backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
bcrypt==5.0.0
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
cryptography==46.0.7
dnspython==2.8.0
ecdsa==0.19.2
email-validator==2.3.0
fastapi==0.136.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
passlib==1.7.4
psycopg2-binary==2.9.12
pyasn1==0.6.3
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.26
requests==2.33.1
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.49
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.45.0
inference-sdk==1.2.5
openai>=1.40.0
google-genai>=1.0.0
transformers>=4.40.0
torch>=2.1.0
safetensors>=0.4.0
Pillow>=10.0.0
